Rename InternalCommandDeduplicateEvents to LuaManagerCommandDeduplicateEvents
