+++ CHANGELOG.md
@@ -19,6 +19,8 @@
    - UICommandWriteToConsole is now UIMethodCollection
    - InternalReconfigured is now InternalCommandReconfigure
    - InternalBootupEvents is now InternalCommandDeduplicateEvents
+   - InternalCommandDeduplicateEvents is now LuaManagerCommandDeduplicateEvents
+ - Plugin renamed: FileTriggerPlugin to LuaManagerPlugin
  - UI: Stores/restores Window position and size
  - Event: Adds TwitchReceivedMessage event (captures all messages, not only commands)
  - Event: Removes TwitchReceivedCommand event as this is already sent as a TwitchReceivedMessage
