+++ Slipstream.csproj
@@ -171,7 +171,7 @@
     <Compile Include="Shared\EventFactory.cs" />
     <Compile Include="Shared\Events\FileMonitor\FileMonitorCommandScan.cs" />
     <Compile Include="Shared\Events\FileMonitor\FileMonitorScanCompleted.cs" />
-    <Compile Include="Shared\Events\Internal\InternalCommandDeduplicateEvents.cs" />
+    <Compile Include="Shared\Events\LuaManager\LuaManagerCommandDeduplicateEvents.cs" />
     <Compile Include="Shared\Events\Internal\InternalCommandReconfigure.cs" />
     <Compile Include="Shared\Events\Internal\InternalCommandPluginStates.cs" />
     <Compile Include="Shared\Events\IRacing\IRacingCommandSendTrackInfo.cs" />
